changed index check case
i am a dumbass who can't count. the reason the test case where index was set to nine worked but nothing else did was the max index of the array was 8. the index i want to check is 7. the reason those sections were deleted was because they contained angle brackets, which were interpreted as html tags.

TODO: separate the array into a separate file, somehow don't hardcode index 7.